Bump kernel to jessie-backports
The stock jessie kernel 3.16 fails to replace the missing disk.  Bumping to
the jessie-backports 4.9 kernel is demonstarted to work as expected.

For this to all function correct we need to drop back to the vanilla jessie64
box and make use of the vagrant-vbguest pluging to install the guest additions
because the stock packages fail to build correctly with the backported kernel.